pyproject.toml
[project]
name = "cv_fneira"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "graphi>=0.1.0",
    "jinja2>=3.1",
    "pypdf>=6.10.2",
    "python-docx>=1.2.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "watchfiles>=1.0",
]
